Fixed CoxTime to work wtith tensor input
